requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
easydict==1.9
graphviz==0.8.4
idna==2.10
imageio==2.9.0
imutils==0.5.3
insightface==0.1.5
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
mxnet==1.6.0
networkx==2.5
numpy==1.19.3
opencv-python==4.4.0.46
Pillow==8.0.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
PyWavelets==1.1.1
requests==2.25.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.12.8
tornado==6.1
tqdm==4.54.1
tzlocal==2.1
urllib3==1.26.2
